requirements.txt

flask
sqlalchemy
python-dateutil
passlib
flask_sqlalchemy
flask_bcrypt

flask_login
flask_mail
flask_wtf

pillow
flask-assets
stripe

xlrd

email_validator
wtforms


certifi==2020.4.5.2
chardet==3.0.4
Click==7.1.2

idna==2.9
itsdangerous==1.1.0

MarkupSafe==1.1.1
python-dotenv==0.13.0
requests==2.24.0

toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1









